Make the input path measurably fast, and prove it with an instrument
The complaint was a felt delay between pressing a macro key and the macro
happening. This adds the instrument that answers that question, then fixes
what the instrument actually showed rather than what seemed likely.

neuron-core/src/latency.rs — always-on, lock-free per-stage histograms
(4 sub-buckets per octave; percentiles quoted at bucket ceilings so a number
is never flattering). Recording costs an Instant::now plus a few relaxed
atomics, so it stays on in release: when a press feels slow, its numbers are
already recorded. press_to_output spans threads via a consumed-once
thread-local, so no signature churn on the action layer.

Measured, then fixed (release, same machine, before -> after):

  * input-thread priority is the big one. Against above-normal contention —
    what a game creates — the pump hop was 7.3ms mean / 24.6ms p99 at default
    priority and 40us / 255us boosted. ~180x. Against merely normal-priority
    load there is no difference at all, because Windows already boosts a
    thread waking from a wait; measuring only that would have concluded the
    boost does nothing. Applied to the pump, both LL-hook pumps and the HID
    readers — deliberately NOT the macro pool, which runs arbitrary user
    macro bodies and measured identical either way.
  * ctx_capture 734us -> 11us mean (p99 5.1ms -> 27us): Context::capture
    probed the foreground app and title twice each, and app.rs now memoizes
    the OpenProcess round trip on the (hwnd, pid) pair.
  * macro_spawn 342us -> 37us mean (p99 2.6ms -> 255us): a fresh OS thread
    per macro fire, paid on the dispatch thread, replaced by a warm bounded
    pool (macros/runner.rs). A full queue refuses honestly instead of
    stalling the pump or silently dropping.
  * shell/run macros 5.7ms -> 5us mean (p99 17.4ms -> 131us): CreateProcess
    ran inline on the dispatch thread, deafening every other binding for its
    duration. Now handed to the pool, but only when actually servicing an
    input edge, since a one-shot CLI would exit before a detached worker ran.
  * Python macro fires did a synchronous options-file read per press.

Two things the measurements killed, recorded so they are not redone: a
hand-rolled high-resolution sleep timer (Rust 1.75+ std already does exactly
that — indistinguishable when compared side by side) and SendInput batching
(the cost is per event, not per call: 10 keystrokes in one call was 1.1x).
SendInput remains ~1.3ms per keystroke and is not ours to fix.

Bugs found while reviewing this work, fixed at the root:

  * a trigger matching several rules ran later rules with the input-path flag
    already consumed by an earlier keystroke, putting CreateProcess back on
    the pump whenever a Key rule preceded a Shell rule. The measurement and
    the scheduling fact are now separate thread-locals.
  * moving the spawn off-thread opened a TOCTOU on the process-spawn arm
    gate; the worker re-checks before spawning.
  * the macro options cache could lose a fresh UI save to a slow reader.
  * spawn_shell_command passed the command line through Rust's MSVC argument
    escaping, which rewrites embedded quotes as \" — cmd.exe does not
    understand that, so every user shell macro with a quoted path silently
    failed. Now raw_arg.
  * hid_usage_for_key folded case in two places, so "F5" returned None while
    "G" worked.
  * three hardcoded 6.283 literals stood in for TAU in the weave renderer.

Reporting: neuron_latency.on in the run root turns capture on with no
restart and no elevation (the app is launched by a scheduled task, where env
vars are unreachable), self-truncating at 4MB. Every report carries the
process's own scheduling posture, because the same machine reads 40us or
7.3ms depending on it. The crash dump gets an allocation-free variant, since
the flight recorder must not call the allocator from a panic handler.

Also folds in the pending licensing sweep (SPDX headers, LICENSES/,
THIRD-PARTY-NOTICES.md, the license-boundary test), the feel-intent and
device-scoped remap modules, the mock transport seam, and the macro-host
sidecar sources.

Whole workspace green: 1406 tests plus doctests, zero rustc warnings, zero
clippy errors.

+++ crates/engram/src/fit.rs
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2026 Woflo Labs
+// SPDX-License-Identifier: LicenseRef-WLCSL-1.0
+// See ../LICENSE.md.
+
 //! The core AR(2) fitting algorithm.
 //!
 //! Five dot products and a division. This is the entire engine.
@@ -286,7 +290,7 @@ mod tests {
     fn fit_pair_damped() {
         let z: Vec<Complex64> = (0..60)
             .map(|i| {
-                let decay = 0.95_f64.powi(i as i32);
+                let decay = 0.95_f64.powi(i);
                 Complex64::new(decay * (2.0 * PI * i as f64 / 12.0).cos(), 0.0)
             })
             .collect();
